+++ Android/.gradle/buildOutputCleanup/cache.properties
@@ -1,0 +1,2 @@
+#Thu Feb 02 14:01:42 CST 2023
+gradle.version=5.4.1
