+++ Android/README.md
@@ -1,2 +1,2 @@
-# ImitationDYShortVideo
-仿抖音上下滑动列表播放短视频解决方案
+# AgoraMultiChannelSwitch
+Agora Multi Channel Switch Sample
