+++ Android/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">ImitationDYShortVideo</string>
+    <string name="app_name">AgoraMultiChannelSwitch</string>
 </resources>
